Initial version of a Product Manager
Initial version for a Product Manager agent as well as a test interface to communicate and ask for the product that we want our agents to develop. New classes: PMAgent.php for the new PM Agent, chat_pm.php to communicate with the agent from the index.php interface, and minor fix in Agent.php so that callLLM function can be called from PMAgent.php.